Navbar working on screen sizes till md
