+++ gatsby-ssr.js
@@ -3,9 +3,14 @@ https://www.gatsbyjs.com/docs/reference/built-in-components/gatsby-head/#editing
 https://www.gatsbyjs.com/docs/reference/config-files/gatsby-ssr/
 */
 
-exports.onRenderBody = ({ setBodyAttributes }, pluginOptions) => {
+exports.onRenderBody = (
+  { setHtmlAttributes, setBodyAttributes },
+  pluginOptions
+) => {
+  setHtmlAttributes({
+    className: "scroll-smooth",
+  })
   setBodyAttributes({
-    className:
-      "bg-slate-100 text-slate-800 max-w-screen-sm md:max-w-screen-md mx-auto",
+    className: "bg-slate-100 text-slate-800",
   })
 }
