+++ src/components/Footer.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 
 const Footer = () => {
   return (
-    <footer className="mx-4 mb-4 text-lg">
+    <footer className="w-full max-w-screen-sm md:max-w-screen-md mx-auto px-4 mb-4 text-lg">
       <p>
         <span className="font-bold">Contact:</span> sahelm.iqbal@niser.ac.in
       </p>
