+++ src/components/Layout.tsx
@@ -12,7 +12,9 @@ const Layout = ({ children }: LayoutProps) => {
     <div className="flex flex-col min-h-screen">
       <div className="grow">
         <Header />
-        <main className="mx-4">{children}</main>
+        <main className="max-w-screen-sm md:max-w-screen-md mx-auto px-4 mt-20">
+          {children}
+        </main>
       </div>
       <Footer />
     </div>
